ffi: remove embedded tor client
Embedded tor client will be available only in pure rust libs. In the bindings isn't possible to choose what feature to enable or disable, and the tor embedded client increases the binary size (i.e., in android ARM64 the `.so` lib increase of ~4MB). Proxy is still supported, so devs can choose to manually embed a tor client and use it with the proxy.

Signed-off-by: Yuki Kishimoto <yukikishimoto@protonmail.com>