Drop zapper support
* Remove `nostr-zapper`
* Remove `Client::zap`

The zapper is too high level. This will also reduce a little bit the binary size of the bindings and the project complexity. In addition, the `Client::zap` method wasn't really configurable (fetch every time user metadata from relays and not allow to choose if fetch from the database, use fixed relays to publish the zap, use fixed timeout, don't allow to pass a LNURL or LN address) and making it configurable would have complicated the library even more. So, remove the zapper and let the devs implement or use other libraries for zapping.

Signed-off-by: Yuki Kishimoto <yukikishimoto@protonmail.com>