Move NIP-07 to a dedicated library
The NIP-07 is browser-only and has browser-specific dependencies. Instead of keeping it under a feature in the `nostr` crate, move it to a dedicated crate. This helps to reduce a little bit the number of dependencies in the `nostr` crate and to remove some browser-specific code.

Pull-Request: https://github.com/rust-nostr/nostr/pull/937
Signed-off-by: Yuki Kishimoto <yukikishimoto@protonmail.com>